requirements.txt
openai>=1.3.0
Pillow>=10.0.0
google-genai
Pyobjc